Move ReviewDate into patterns directory
