+++ src/index.ts
@@ -42,7 +42,7 @@ export { default as ListPanel } from './components/list-panel';
 export { default as NavAZ } from './components/nav-a-z';
 export { default as Pagination } from './components/pagination';
 export { default as Radios } from './components/radios';
-export { default as ReviewDate } from './components/review-date';
+export { default as ReviewDate } from './patterns/review-date';
 export { default as Select } from './components/select';
 export { default as SkipLink } from './components/skip-link';
 export { default as SummaryList } from './components/summary-list';
