+++ stories/Patterns/ReviewDate.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ReviewDate } from '../src';
+import { ReviewDate } from '../../src';
 
 export const Standard = (): JSX.Element => (
   <ReviewDate lastReviewed="12 Feburary 2016" nextReview="1 Feburary 2019" />
@@ -10,6 +10,6 @@ export const JustLastReview = (): JSX.Element => <ReviewDate lastReviewed="12 Fe
 export const JustNextReview = (): JSX.Element => <ReviewDate nextReview="1 Feburary 2019" />;
 
 export default {
-  title: 'Components/ReviewDate',
+  title: 'Patterns/ReviewDate',
   component: ReviewDate,
 };
